+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Renamed "You are not allowed to send messages" to "You can not send messages" (sounds nicer ❤)
 * Chatlist crashed randomly
 * Sending of whitespaces only was allowed
++ Fixed wrong message bubble height
 
 ### 7.0:
 ##### New features